lambdas/data-upload-to-mp/requirements.txt
boto3==1.9.74
pandas==0.23.4
requests==2.21.0
numpy==1.15.4
matplotlib==3.0.2
botocore==1.12.74
python_dateutil==2.8.0
